requirements_dev.txt
bumpversion==0.5.3
Click==7.0
coverage==4.5.1
flake8==3.5.0
wheel==0.32.1
watchdog==0.9.0
tox==3.5.2
python-dotenv==0.10.3
pyyaml==5.1.1 
twine==1.12.1

pytest==3.8.2
pytest-runner==4.2
